pci: recoverable ECAM config reads on aarch64 (Apple aborts unlinked buses)
/wifi power faulted with a data abort (ESR 0x96000010, FAR 0x690200000 = ECAM +
bus2<<20) after the /wifi reset power-cycle/PERST work: Apple's APCIE ECAM
external-aborts a config read to an unlinked secondary bus instead of returning
all-ones, so `pci::for_each` walking 0..=bus_end (4 on t8112) FATALs the moment
it reads bus 2. Before the power-cycle that bus read back clean; the reset left
the fabric aborting it.

Route pci::read32 through the recoverable probe on aarch64 (non-test) so an
absent/unlinked bus reads as 0xffffffff, matching standard PCI behaviour — the
same recovery every Apple MMIO/config path already uses. x86 keeps the raw
volatile read (cfg-gated), so it's byte-identical there. Both arches build; unit
suite green.

+++ kernel/src/drivers/wifi/iwl/mod.rs
@@ -349,8 +349,8 @@ pub fn scan() -> Result<alloc::vec::Vec<crate::drivers::wifi::proto::BssInfo>, &
         .map(|entry| crate::drivers::wifi::proto::BssInfo {
             ssid: entry.bss.ssid,
             bssid: entry.bss.bssid,
-            channel: entry.bss.channel,
-            rssi: entry.rssi.unwrap_or(-127),
+            channel: entry.bss.channel as u16,
+            rssi: entry.rssi.unwrap_or(-127) as i16,
             privacy: entry.bss.privacy,
         })
         .collect())
